fix: rename modal
