+++ src/shared/ui/forms/forgotPassword/ForgotPasswordForm.tsx
@@ -7,7 +7,7 @@ import { usePasswordRecoveryMutation } from '@/app/api/inctagramApi'
 import { emailSchema } from '@/shared/model/schemas/schemas'
 import { Button } from '@/shared/ui/button'
 import { InputControl } from '@/shared/ui/inputControl'
-import { Modal } from '@/shared/ui/modal/modal'
+import { Modal } from '@/shared/ui/modal/renamedModal'
 import Recaptcha from '@/shared/ui/recaptcha/Recaptcha'
 import { Typography } from '@/shared/ui/typography'
 import { zodResolver } from '@hookform/resolvers/zod'
