+++ src/shared/ui/modal/index.ts
@@ -1,1 +1,1 @@
-export * from './modal'
+export * from './renamedModal'
