+++ src/shared/ui/modal/modal.stories.tsx
@@ -4,7 +4,7 @@ import { useState } from 'react'
 
 import { Button } from '../button'
 import { Typography } from '../typography'
-import { Modal } from './modal'
+import { Modal } from './renamedModal'
 
 const meta = {
   argTypes: {
